FastApiService/requirements.txt
fastapi
pydantic
loguru  # A more modern logging library; recommended for logging
openai==0.28
python-dotenv
requests
beautifulsoup4